requirements.txt
allure-pytest==2.7.0
allure-python-commons==2.7.0
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
Faker==6.6.3
idna==2.8
importlib-metadata==3.7.3
more-itertools==8.7.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.0.1
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
